Enable PSR-4 for tests.
