+++ composer.json
@@ -27,5 +27,10 @@
         "psr-4": {
             "MessageBird\\": "src/MessageBird/"
         }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Tests\\": "tests/"
+        }
     }
 }
