--- tests/integration/balance/BalanceTest.php
@@ -1,16 +1,0 @@
-<?php
-class BalanceTest extends BaseTest
-{
-    protected function setUp()
-    {
-        parent::setUp();
-        $this->client = new \MessageBird\Client('YOUR_ACCESS_KEY', $this->mockClient);
-    }
-
-    public function testReadBalance()
-    {
-        $this->expectException(\MessageBird\Exceptions\ServerException::class);
-        $this->mockClient->expects($this->once())->method('performHttpRequest')->with("GET", 'balance', null, null);
-        $this->client->balance->read();
-    }
-}
